requirement.txt
pyspark==3.5.3
kafka-python==2.0.2
apache-airflow==2.10.2
